requirements.txt
flask
pillow
numpy
scikit-image
matplotlib
scipy
scikit-learn